+++ src/api/consumer/lib/bind.js
@@ -1,0 +1,7 @@
+// @flow
+import { loadConsumer, Consumer } from '../../../consumer';
+
+export default async function bindAction() {
+  const consumer: Consumer = await loadConsumer();
+  return consumer.bindAll();
+}