+++ src/api/consumer/lib/import.js
@@ -8,6 +8,7 @@ import { Consumer, loadConsumer } from '../../../consumer';
 import loader from '../../../cli/loader';
 import { BEFORE_IMPORT_ENVIRONMENT } from '../../../cli/loader/loader-messages';
 import { flattenDependencies } from '../../../scope/flatten-dependencies';
+import { COMPONENT_ORIGINS } from '../../../constants';
 
 const key = R.compose(R.head, R.keys);
 
@@ -55,9 +56,16 @@ export default async function importAction({
   const cache = false;
   const { dependencies, envDependencies } = await consumer.import(ids, verbose, environment, cache, prefix);
   const bitIds = dependencies.map(R.path(['component', 'id']));
-  if (!R.isEmpty(ids)) {
+  const bitMap = await consumer.getBitMap();
+  const notAuthored = (bitId) => {
+    const componentMap = bitMap.getComponent(bitId);
+    return componentMap && componentMap.origin !== COMPONENT_ORIGINS.AUTHORED;
+  };
+  const notAuthoredBitIds = R.filter(notAuthored, bitIds);
+  console.log(notAuthoredBitIds);
+  if (!R.isEmpty(notAuthoredBitIds)) {
     // not needed when importing from bit.json/bit.map
-    await consumer.bitJson.addDependencies(bitIds).write({ bitDir: consumer.getPath() });
+    await consumer.bitJson.addDependencies(notAuthoredBitIds).write({ bitDir: consumer.getPath() });
   }
 
   const warnings = await warnForPackageDependencies({