+++ src/consumer/component/consumer-component.js
@@ -491,8 +491,11 @@ export default class Component {
 
           return this.specsResults;
         } catch (err) {
-          if (verbose) throw err;
-          throw new ComponentSpecsFailed();
+          if (rejectOnFailure) {
+            if (verbose) throw err;
+            throw new ComponentSpecsFailed();
+          }
+          return this.specsResults;
         }
       };