+++ src/consumer/consumer.js
@@ -16,6 +16,8 @@ import { Scope, ComponentWithDependencies } from '../scope';
 import loader from '../cli/loader';
 import { BEFORE_IMPORT_ACTION } from '../cli/loader/loader-messages';
 import BitMap from './bit-map/bit-map';
+import ComponentMap from './bit-map/component-map';
+import type { BitMapComponents } from './bit-map/bit-map';
 import logger from '../logger/logger';
 import DirStructure from './dir-structure/dir-structure';
 import { getLatestVersionNumber } from '../utils';
@@ -86,7 +88,7 @@ export default class Consumer {
    * @returns {boolean} true if the driver exists, false otherwise
    * @memberof Consumer
    */
-  warnForMissingDriver(msg): boolean {
+  warnForMissingDriver(msg: string): boolean {
     try {
       this.driver.getDriver(false);
       return true;
@@ -282,11 +284,11 @@ export default class Consumer {
   }
 
   /**
-   * write the components into '/components' dir (or according to the bit.map) and its
-   * dependencies nested inside the component directory and under 'dependencies' dir.
+   * write the components into '/components' dir (or according to the bit.map) and its dependencies in the
+   * '/components/.dependencies' dir. Both directories are configurable in bit.json
    * For example: global/a has a dependency my-scope/global/b@1. The directories will be:
-   * project/root/component/global/a/impl.js
-   * project/root/component/global/a/dependency/global/b/my-scope/1/impl.js
+   * project/root/components/global/a/impl.js
+   * project/root/components/.dependencies/global/b/my-scope/1/impl.js
    *
    * In case there are some same dependencies shared between the components, it makes sure to
    * write them only once.
@@ -298,9 +300,8 @@ export default class Consumer {
   ): Promise<Component[]> {
     const bitMap: BitMap = await this.getBitMap();
     const dependenciesIdsCache = [];
-
     const writeComponentsP = componentDependencies.map((componentWithDeps) => {
-      const bitDir = writeToPath || this.composeBitPath(componentWithDeps.component.id);
+      const bitDir = writeToPath || this.composeComponentPath(componentWithDeps.component.id);
       componentWithDeps.component.writtenPath = bitDir;
       return componentWithDeps.component.write({
         bitDir,
@@ -331,7 +332,7 @@ export default class Consumer {
           return Promise.resolve();
         }
 
-        const depBitPath = path.join(componentWithDeps.component.writtenPath, DEPENDENCIES_DIR, dep.id.toFullPath());
+        const depBitPath = this.composeDependencyPath(dep.id);
         dep.writtenPath = depBitPath;
         dependenciesIdsCache[dependencyId] = depBitPath;
         return dep
@@ -351,6 +352,16 @@ export default class Consumer {
     });
     const writtenDependencies = await Promise.all(allDependenciesP);
 
+    if (writeToPath) {
+      componentDependencies.forEach((componentWithDeps) => {
+        const relativeWrittenPath = this.getPathRelativeToConsumer(componentWithDeps.component.writtenPath);
+        if (relativeWrittenPath !== writeToPath) {
+          const component = componentWithDeps.component;
+          this.moveExistingComponent(bitMap, component, relativeWrittenPath, writeToPath);
+        }
+      });
+    }
+
     await linkGenerator.writeDependencyLinks(componentDependencies, bitMap, this.getPath());
     await Promise.all(
       componentDependencies.map(componentWithDependencies =>
@@ -358,14 +369,11 @@ export default class Consumer {
       )
     );
 
-    if (writeToPath) {
-      componentDependencies.forEach((componentWithDeps) => {
-        if (componentWithDeps.component.writtenPath !== writeToPath) {
-          const component = componentWithDeps.component;
-          this.moveExistingComponent(bitMap, component, component.writtenPath, writeToPath);
-        }
-      });
-    }
+    const componentsToBind = {};
+    writtenComponents.forEach((component) => {
+      componentsToBind[component.id.toString()] = bitMap.getComponent(component.id);
+    });
+    this.bindComponents(componentsToBind);
 
     await bitMap.write();
     return [...writtenComponents, ...writtenDependencies];
@@ -451,7 +459,7 @@ export default class Consumer {
     logger.debug(`committing the following components: ${ids.join(', ')}`);
     const componentsIds = ids.map(componentId => BitId.parse(componentId));
     const components = await this.loadComponents(componentsIds);
-    // Run over the components to check if there is missing depenedencies
+    // Run over the components to check if there is missing dependencies
     // If there is at least one we won't commit anything
     const componentsWithMissingDeps = components.filter((component) => {
       return Boolean(component.missingDependencies);
@@ -470,6 +478,40 @@ export default class Consumer {
     return components;
   }
 
+  bindComponents(bitMapComponents: BitMapComponents): Object[] {
+    return Object.keys(bitMapComponents).map((componentId) => {
+      logger.debug(`binding component: ${componentId}`);
+      const componentMap: ComponentMap = bitMapComponents[componentId];
+      const parsedId = BitId.parse(componentId);
+      if (componentMap.rootDir) {
+        const target = path.join(this.getPath(), componentMap.rootDir);
+        const relativeLinkPath = path.join('node_modules', 'bit', parsedId.box, parsedId.name);
+        const linkPath = path.join(this.getPath(), relativeLinkPath);
+        const relativeTarget = path.relative(path.dirname(linkPath), target);
+        fs.removeSync(linkPath); // in case a component has been moved
+        fs.ensureSymlinkSync(relativeTarget, linkPath, 'dir');
+        return { id: componentId, bound: [{ from: componentMap.rootDir, to: relativeLinkPath }] };
+      }
+      // const filesToBind = {
+      //   name: parsedId.name,
+      //   namespace: parsedId.box,
+      //   files: componentMap.getFilesRelativeToConsumer(),
+      //   mainFile: componentMap.getMainFileRelativeToConsumer()
+      // };
+      // todo: implement
+      return { id: componentId, bound: [] };
+    });
+  }
+
+  async bindAll() {
+    const bitMap = await this.getBitMap();
+    const nonNestedFilter = component => component.origin !== COMPONENT_ORIGINS.NESTED;
+    const componentsToBind = R.filter(nonNestedFilter, bitMap.getAllComponents());
+    if (R.isEmpty(componentsToBind)) throw new Error('nothing to bind');
+    fs.removeSync(path.join(this.getPath(), 'node_modules', 'bit')); // todo: move to bit-javascript
+    return this.bindComponents(componentsToBind);
+  }
+
   composeRelativeBitPath(bitId: BitId): string {
     const { staticParts, dynamicParts } = this.dirStructure.componentsDirStructure;
     const dynamicDirs = dynamicParts.map(part => bitId[part]);
@@ -477,10 +519,15 @@ export default class Consumer {
     return path.join(...addToPath);
   }
 
-  composeBitPath(bitId: BitId): string {
+  composeComponentPath(bitId: BitId): string {
     const addToPath = [this.getPath(), this.composeRelativeBitPath(bitId)];
     logger.debug(`component dir path: ${addToPath.join('/')}`);
     return path.join(...addToPath);
+  }
+
+  composeDependencyPath(bitId: BitId): string {
+    const dependenciesDir = this.dirStructure.dependenciesDirStructure;
+    return path.join(this.getPath(), dependenciesDir, bitId.toFullPath());
   }
 
   runComponentSpecs(id: BitId, verbose: boolean = false): Promise<?any> {