+++ CHANGELOG.md
@@ -7,10 +7,15 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [unreleased]
 
+- support requiring imported components using `require('bit/namespace/name')` syntax
+- save all dependencies on one configurable directory (components/.dependencies by default) 
 - move a re-imported component to a new location when `bit import --prefix` is used
 - fix commit and export issues when dealing with more than 500 components
 - support move command for moving files/directories of a component to a new location
 - fix export of large amount of data
+- fix bug with commit --force when tests throws exception
+- fix bug - when you import authored component (and there is newer version) it duplicate it in the .bit.map.json
+- fix bug - when you import authored component it was added to bit.json dependencies
 - create package.json for imported components
 
 ## [0.10.7] - 2017-09-07